Tried to fix Email Issue & Created different file for state&district data
