+++ src/auth/Register.jsx
@@ -1,6 +1,9 @@
 import { useState, useEffect } from "react";
 import logo from "../assets/logo1.png";
 import emailjs from "emailjs-com";
+import { fetchSignInMethodsForEmail } from "firebase/auth";
+
+
 
 import {
   FaEye,
@@ -21,795 +24,10 @@ import {
 import { ref, set, get } from "firebase/database";
 import { toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
+import statesWithDistricts from "../components/StatesWithDistricts.jsx";
 
-// ✅ Hardcoded States & Districts
-const statesWithDistricts = {
-  Maharashtra: [
-    "Ahmednagar",
-    "Akola",
-    "Amravati",
-    "Aurangabad",
-    "Beed",
-    "Bhandara",
-    "Buldhana",
-    "Chandrapur",
-    "Dhule",
-    "Gadchiroli",
-    "Gondia",
-    "Hingoli",
-    "Jalgaon",
-    "Jalna",
-    "Kolhapur",
-    "Latur",
-    "Mumbai City",
-    "Mumbai Suburban",
-    "Nagpur",
-    "Nanded",
-    "Nandurbar",
-    "Nashik",
-    "Osmanabad",
-    "Palghar",
-    "Parbhani",
-    "Pune",
-    "Raigad",
-    "Ratnagiri",
-    "Sangli",
-    "Satara",
-    "Sindhudurg",
-    "Solapur",
-    "Thane",
-    "Wardha",
-    "Washim",
-    "Yavatmal",
-  ],
-  Gujarat: [
-    "Ahmedabad",
-    "Amreli",
-    "Anand",
-    "Aravalli",
-    "Banaskantha",
-    "Bharuch",
-    "Bhavnagar",
-    "Botad",
-    "Chhota Udaipur",
-    "Dahod",
-    "Dang",
-    "Devbhoomi Dwarka",
-    "Gandhinagar",
-    "Gir Somnath",
-    "Jamnagar",
-    "Junagadh",
-    "Kheda",
-    "Kutch",
-    "Mahisagar",
-    "Mehsana",
-    "Morbi",
-    "Narmada",
-    "Navsari",
-    "Panchmahal",
-    "Patan",
-    "Porbandar",
-    "Rajkot",
-    "Sabarkantha",
-    "Surat",
-    "Surendranagar",
-    "Tapi",
-    "Vadodara",
-    "Valsad",
-  ],
-  Karnataka: [
-    "Bagalkot",
-    "Ballari",
-    "Belagavi",
-    "Bengaluru Rural",
-    "Bengaluru Urban",
-    "Bidar",
-    "Chamarajanagar",
-    "Chikkaballapur",
-    "Chikkamagaluru",
-    "Chitradurga",
-    "Dakshina Kannada",
-    "Davangere",
-    "Dharwad",
-    "Gadag",
-    "Hassan",
-    "Haveri",
-    "Kalaburagi",
-    "Kodagu",
-    "Kolar",
-    "Koppal",
-    "Mandya",
-    "Mysuru",
-    "Raichur",
-    "Ramanagara",
-    "Shivamogga",
-    "Tumakuru",
-    "Udupi",
-    "Uttara Kannada",
-    "Vijayapura",
-    "Yadgir",
-    "Vijayanagara",
-  ],
-  AndhraPradesh: [
-    "Anantapur",
-    "Chittoor",
-    "East Godavari",
-    "Guntur",
-    "Krishna",
-    "Kurnool",
-    "Prakasam",
-    "SPS Nellore",
-    "Srikakulam",
-    "Visakhapatnam",
-    "Vizianagaram",
-    "West Godavari",
-    "YSR Kadapa",
-  ],
-  ArunachalPradesh: [
-    "Tawang",
-    "West Kameng",
-    "East Kameng",
-    "Papum Pare",
-    "Kurung Kumey",
-    "Kra Daadi",
-    "Lower Subansiri",
-    "Upper Subansiri",
-    "West Siang",
-    "East Siang",
-    "Siang",
-    "Upper Siang",
-    "Lower Siang",
-    "Lower Dibang Valley",
-    "Dibang Valley",
-    "Anjaw",
-    "Lohit",
-    "Namsai",
-    "Changlang",
-    "Tirap",
-    "Longding",
-    "Kamle",
-    "Leparada",
-    "Shi Yomi",
-    "Pakke-Kessang",
-    "Lepa Rada",
-  ],
-  Assam: [
-    "Baksa",
-    "Barpeta",
-    "Biswanath",
-    "Bongaigaon",
-    "Cachar",
-    "Charaideo",
-    "Chirang",
-    "Darrang",
-    "Dhemaji",
-    "Dhubri",
-    "Dibrugarh",
-    "Goalpara",
-    "Golaghat",
-    "Hailakandi",
-    "Hojai",
-    "Jorhat",
-    "Kamrup",
-    "Kamrup Metropolitan",
-    "Karbi Anglong",
-    "Karimganj",
-    "Kokrajhar",
-    "Lakhimpur",
-    "Majuli",
-    "Morigaon",
-    "Nagaon",
-    "Nalbari",
-    "Sivasagar",
-    "Sonitpur",
-    "South Salmara-Mankachar",
-    "Tinsukia",
-    "Udalguri",
-    "West Karbi Anglong",
-  ],
-  Bihar: [
-    "Araria",
-    "Arwal",
-    "Aurangabad",
-    "Banka",
-    "Begusarai",
-    "Bhagalpur",
-    "Bhojpur",
-    "Buxar",
-    "Darbhanga",
-    "East Champaran",
-    "Gaya",
-    "Gopalganj",
-    "Jamui",
-    "Jehanabad",
-    "Kaimur",
-    "Katihar",
-    "Khagaria",
-    "Kishanganj",
-    "Lakhisarai",
-    "Madhepura",
-    "Madhubani",
-    "Munger",
-    "Muzaffarpur",
-    "Nalanda",
-    "Nawada",
-    "Patna",
-    "Purnia",
-    "Rohtas",
-    "Saharsa",
-    "Samastipur",
-    "Saran",
-    "Sheikhpura",
-    "Sheohar",
-    "Sitamarhi",
-    "Siwan",
-    "Supaul",
-    "Vaishali",
-    "West Champaran",
-  ],
-  Chhattisgarh: [
-    "Balod",
-    "Baloda Bazar",
-    "Balrampur",
-    "Bastar",
-    "Bemetara",
-    "Bijapur",
-    "Bilaspur",
-    "Dantewada",
-    "Dhamtari",
-    "Durg",
-    "Gariaband",
-    "Gaurela-Pendra-Marwahi",
-    "Janjgir-Champa",
-    "Jashpur",
-    "Kabirdham",
-    "Kanker",
-    "Kondagaon",
-    "Korba",
-    "Koriya",
-    "Mahasamund",
-    "Mungeli",
-    "Narayanpur",
-    "Raigarh",
-    "Raipur",
-    "Rajnandgaon",
-    "Sukma",
-    "Surajpur",
-    "Surguja",
-  ],
-  Goa: ["North Goa", "South Goa"],
-  Haryana: [
-    "Ambala",
-    "Bhiwani",
-    "Charkhi Dadri",
-    "Faridabad",
-    "Fatehabad",
-    "Gurugram",
-    "Hisar",
-    "Jhajjar",
-    "Jind",
-    "Kaithal",
-    "Karnal",
-    "Kurukshetra",
-    "Mahendragarh",
-    "Nuh",
-    "Palwal",
-    "Panchkula",
-    "Panipat",
-    "Rewari",
-    "Rohtak",
-    "Sirsa",
-    "Sonipat",
-    "Yamunanagar",
-  ],
-  HimachalPradesh: [
-    "Bilaspur",
-    "Chamba",
-    "Hamirpur",
-    "Kangra",
-    "Kinnaur",
-    "Kullu",
-    "Lahaul and Spiti",
-    "Mandi",
-    "Shimla",
-    "Sirmaur",
-    "Solan",
-    "Una",
-  ],
-  Jharkhand: [
-    "Bokaro",
-    "Chatra",
-    "Deoghar",
-    "Dhanbad",
-    "Dumka",
-    "East Singhbhum",
-    "Garhwa",
-    "Giridih",
-    "Godda",
-    "Gumla",
-    "Hazaribagh",
-    "Jamtara",
-    "Khunti",
-    "Koderma",
-    "Latehar",
-    "Lohardaga",
-    "Pakur",
-    "Palamu",
-    "Ramgarh",
-    "Ranchi",
-    "Sahebganj",
-    "Seraikela Kharsawan",
-    "Simdega",
-    "West Singhbhum",
-  ],
-  Kerala: [
-    "Alappuzha",
-    "Ernakulam",
-    "Idukki",
-    "Kannur",
-    "Kasaragod",
-    "Kollam",
-    "Kottayam",
-    "Kozhikode",
-    "Malappuram",
-    "Palakkad",
-    "Pathanamthitta",
-    "Thiruvananthapuram",
-    "Thrissur",
-    "Wayanad",
-  ],
-  MadhyaPradesh: [
-    "Agar Malwa",
-    "Alirajpur",
-    "Anuppur",
-    "Ashoknagar",
-    "Balaghat",
-    "Barwani",
-    "Betul",
-    "Bhind",
-    "Bhopal",
-    "Burhanpur",
-    "Chhatarpur",
-    "Chhindwara",
-    "Damoh",
-    "Datia",
-    "Dewas",
-    "Dhar",
-    "Dindori",
-    "Guna",
-    "Gwalior",
-    "Harda",
-    "Hoshangabad",
-    "Indore",
-    "Jabalpur",
-    "Jhabua",
-    "Katni",
-    "Khandwa",
-    "Khargone",
-    "Mandla",
-    "Mandsaur",
-    "Morena",
-    "Narsinghpur",
-    "Neemuch",
-    "Panna",
-    "Raisen",
-    "Rajgarh",
-    "Ratlam",
-    "Rewa",
-    "Sagar",
-    "Satna",
-    "Sehore",
-    "Seoni",
-    "Shahdol",
-    "Shajapur",
-    "Sheopur",
-    "Shivpuri",
-    "Sidhi",
-    "Singrauli",
-    "Tikamgarh",
-    "Ujjain",
-    "Umaria",
-    "Vidisha",
-  ],
-  Manipur: [
-    "Bishnupur",
-    "Chandel",
-    "Churachandpur",
-    "Imphal East",
-    "Imphal West",
-    "Jiribam",
-    "Kakching",
-    "Kamjong",
-    "Kangpokpi",
-    "Noney",
-    "Pherzawl",
-    "Senapati",
-    "Tamenglong",
-    "Tengnoupal",
-    "Thoubal",
-    "Ukhrul",
-  ],
-  Meghalaya: [
-    "East Garo Hills",
-    "East Jaintia Hills",
-    "East Khasi Hills",
-    "North Garo Hills",
-    "Ri Bhoi",
-    "South Garo Hills",
-    "South West Garo Hills",
-    "South West Khasi Hills",
-    "West Garo Hills",
-    "West Jaintia Hills",
-    "West Khasi Hills",
-  ],
-  Mizoram: [
-    "Aizawl",
-    "Champhai",
-    "Hnahthial",
-    "Khawzawl",
-    "Kolasib",
-    "Lawngtlai",
-    "Lunglei",
-    "Mamit",
-    "Saiha",
-    "Serchhip",
-    "Saitual",
-  ],
-  Nagaland: [
-    "Chümoukedima",
-    "Dimapur",
-    "Kiphire",
-    "Kohima",
-    "Longleng",
-    "Mokokchung",
-    "Mon",
-    "Niuland",
-    "Noklak",
-    "Peren",
-    "Phek",
-    "Shamator",
-    "Tseminyu",
-    "Tuensang",
-    "Wokha",
-    "Zünheboto",
-  ],
-  Odisha: [
-    "Angul",
-    "Balangir",
-    "Balasore",
-    "Bargarh",
-    "Bhadrak",
-    "Boudh",
-    "Cuttack",
-    "Deogarh",
-    "Dhenkanal",
-    "Gajapati",
-    "Ganjam",
-    "Jagatsinghpur",
-    "Jajpur",
-    "Jharsuguda",
-    "Kalahandi",
-    "Kandhamal",
-    "Kendrapara",
-    "Kendujhar",
-    "Khordha",
-    "Koraput",
-    "Malkangiri",
-    "Mayurbhanj",
-    "Nabarangpur",
-    "Nayagarh",
-    "Nuapada",
-    "Puri",
-    "Rayagada",
-    "Sambalpur",
-    "Subarnapur",
-    "Sundargarh",
-  ],
-  Punjab: [
-    "Amritsar",
-    "Barnala",
-    "Bathinda",
-    "Faridkot",
-    "Fatehgarh Sahib",
-    "Fazilka",
-    "Ferozepur",
-    "Gurdaspur",
-    "Hoshiarpur",
-    "Jalandhar",
-    "Kapurthala",
-    "Ludhiana",
-    "Malerkotla",
-    "Mansa",
-    "Moga",
-    "Mohali",
-    "Muktsar",
-    "Pathankot",
-    "Patiala",
-    "Rupnagar",
-    "Sangrur",
-    "SAS Nagar",
-    "Shahid Bhagat Singh Nagar",
-    "Sri Muktsar Sahib",
-    "Tarn Taran",
-  ],
-  Rajasthan: [
-    "Ajmer",
-    "Alwar",
-    "Banswara",
-    "Baran",
-    "Barmer",
-    "Bharatpur",
-    "Bhilwara",
-    "Bikaner",
-    "Bundi",
-    "Chittorgarh",
-    "Churu",
-    "Dausa",
-    "Dholpur",
-    "Dungarpur",
-    "Hanumangarh",
-    "Jaipur",
-    "Jaisalmer",
-    "Jalore",
-    "Jhalawar",
-    "Jhunjhunu",
-    "Jodhpur",
-    "Karauli",
-    "Kota",
-    "Nagaur",
-    "Pali",
-    "Pratapgarh",
-    "Rajsamand",
-    "Sawai Madhopur",
-    "Sikar",
-    "Sirohi",
-    "Sri Ganganagar",
-    "Tonk",
-    "Udaipur",
-  ],
-  Sikkim: ["East Sikkim", "West Sikkim", "North Sikkim", "South Sikkim"],
-  TamilNadu: [
-    "Ariyalur",
-    "Chengalpattu",
-    "Chennai",
-    "Coimbatore",
-    "Cuddalore",
-    "Dharmapuri",
-    "Dindigul",
-    "Erode",
-    "Kallakurichi",
-    "Kancheepuram",
-    "Karur",
-    "Krishnagiri",
-    "Madurai",
-    "Mayiladuthurai",
-    "Nagapattinam",
-    "Namakkal",
-    "Nilgiris",
-    "Perambalur",
-    "Pudukkottai",
-    "Ramanathapuram",
-    "Ranipet",
-    "Salem",
-    "Sivaganga",
-    "Tenkasi",
-    "Thanjavur",
-    "Theni",
-    "Thoothukudi",
-    "Tiruchirappalli",
-    "Tirunelveli",
-    "Tirupattur",
-    "Tiruppur",
-    "Tiruvallur",
-    "Tiruvannamalai",
-    "Tiruvarur",
-    "Vellore",
-    "Viluppuram",
-    "Virudhunagar",
-  ],
-  Telangana: [
-    "Adilabad",
-    "Bhadradri Kothagudem",
-    "Hanumakonda",
-    "Hyderabad",
-    "Jagitial",
-    "Jangaon",
-    "Jayashankar Bhupalpally",
-    "Jogulamba Gadwal",
-    "Kamareddy",
-    "Karimnagar",
-    "Khammam",
-    "Komaram Bheem Asifabad",
-    "Mahabubabad",
-    "Mahabubnagar",
-    "Mancherial",
-    "Medak",
-    "Medchal–Malkajgiri",
-    "Mulugu",
-    "Nagarkurnool",
-    "Nalgonda",
-    "Nirmal",
-    "Nizamabad",
-    "Peddapalli",
-    "Rajanna Sircilla",
-    "Rangareddy",
-    "Sangareddy",
-    "Siddipet",
-    "Suryapet",
-    "Vikarabad",
-    "Wanaparthy",
-    "Warangal",
-    "Yadadri Bhuvanagiri",
-  ],
-  Tripura: [
-    "Dhalai",
-    "Gomati",
-    "Khowai",
-    "North Tripura",
-    "Sepahijala",
-    "South Tripura",
-    "Unakoti",
-    "West Tripura",
-  ],
-  UttarPradesh: [
-    "Agra",
-    "Aligarh",
-    "Ambedkar Nagar",
-    "Amethi",
-    "Amroha",
-    "Auraiya",
-    "Ayodhya",
-    "Azamgarh",
-    "Baghpat",
-    "Bahraich",
-    "Ballia",
-    "Balrampur",
-    "Banda",
-    "Barabanki",
-    "Bareilly",
-    "Basti",
-    "Bhadohi",
-    "Bijnor",
-    "Budaun",
-    "Bulandshahr",
-    "Chandauli",
-    "Chitrakoot",
-    "Deoria",
-    "Etah",
-    "Etawah",
-    "Farrukhabad",
-    "Fatehpur",
-    "Firozabad",
-    "Gautam Buddh Nagar",
-    "Ghaziabad",
-    "Ghazipur",
-    "Gonda",
-    "Gorakhpur",
-    "Hamirpur",
-    "Hapur",
-    "Hardoi",
-    "Hathras",
-    "Jalaun",
-    "Jaunpur",
-    "Jhansi",
-    "Kannauj",
-    "Kanpur Dehat",
-    "Kanpur Nagar",
-    "Kasganj",
-    "Kaushambi",
-    "Kheri",
-    "Kushinagar",
-    "Lalitpur",
-    "Lucknow",
-    "Maharajganj",
-    "Mahoba",
-    "Mainpuri",
-    "Mathura",
-    "Mau",
-    "Meerut",
-    "Mirzapur",
-    "Moradabad",
-    "Muzaffarnagar",
-    "Pilibhit",
-    "Pratapgarh",
-    "Prayagraj",
-    "Raebareli",
-    "Rampur",
-    "Saharanpur",
-    "Sambhal",
-    "Sant Kabir Nagar",
-    "Shahjahanpur",
-    "Shamli",
-    "Shrawasti",
-    "Siddharthnagar",
-    "Sitapur",
-    "Sonbhadra",
-    "Sultanpur",
-    "Unnao",
-    "Varanasi",
-  ],
-  Uttarakhand: [
-    "Almora",
-    "Bageshwar",
-    "Chamoli",
-    "Champawat",
-    "Dehradun",
-    "Haridwar",
-    "Nainital",
-    "Pauri Garhwal",
-    "Pithoragarh",
-    "Rudraprayag",
-    "Tehri Garhwal",
-    "Udham Singh Nagar",
-    "Uttarkashi",
-  ],
-  WestBengal: [
-    "Alipurduar",
-    "Bankura",
-    "Birbhum",
-    "Cooch Behar",
-    "Dakshin Dinajpur",
-    "Darjeeling",
-    "Hooghly",
-    "Howrah",
-    "Jalpaiguri",
-    "Jhargram",
-    "Kalimpong",
-    "Kolkata",
-    "Malda",
-    "Murshidabad",
-    "Nadia",
-    "North 24 Parganas",
-    "Paschim Bardhaman",
-    "Paschim Medinipur",
-    "Purba Bardhaman",
-    "Purba Medinipur",
-    "Purulia",
-    "South 24 Parganas",
-    "Uttar Dinajpur",
-  ],
-  AndamanNicobar: ["Nicobar", "North and Middle Andaman", "South Andaman"],
-  Chandigarh: ["Chandigarh"],
-  DadraNagarHaveliDamanDiu: ["Dadra and Nagar Haveli", "Daman", "Diu"],
-  Lakshadweep: ["Lakshadweep"],
-  Delhi: [
-    "Central Delhi",
-    "East Delhi",
-    "New Delhi",
-    "North Delhi",
-    "North East Delhi",
-    "North West Delhi",
-    "Shahdara",
-    "South Delhi",
-    "South East Delhi",
-    "South West Delhi",
-    "West Delhi",
-  ],
-  Puducherry: ["Karaikal", "Mahe", "Puducherry", "Yanam"],
-  Ladakh: ["Kargil", "Leh"],
-  JammuKashmir: [
-    "Anantnag",
-    "Bandipora",
-    "Baramulla",
-    "Budgam",
-    "Doda",
-    "Ganderbal",
-    "Jammu",
-    "Kathua",
-    "Kishtwar",
-    "Kulgam",
-    "Kupwara",
-    "Poonch",
-    "Pulwama",
-    "Rajouri",
-    "Ramban",
-    "Reasi",
-    "Samba",
-    "Shopian",
-    "Srinagar",
-    "Udhampur",
-  ],
-};
+
+
 
 const Register = ({ onClose, onSwitchToLogin }) => {
   const [showPassword, setShowPassword] = useState(false);
@@ -863,8 +81,8 @@ const Register = ({ onClose, onSwitchToLogin }) => {
     const strongRegex = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&]).{8,}$/;
 
     if (weakRegex.test(pwd)) setPasswordStrength("Weak");
-    else if (mediumRegex.test(pwd)) setPasswordStrength("Medium");
     else if (strongRegex.test(pwd)) setPasswordStrength("Strong");
+    else if (mediumRegex.test(pwd)) setPasswordStrength("Medium");
   }, [formData.password]);
 
   // ✅ Password Validation (strict)
@@ -896,32 +114,45 @@ const Register = ({ onClose, onSwitchToLogin }) => {
     }
 
     setLoading(true);
+
     try {
+      // 🚨 Double check email before OTP
+       const methods = await fetchSignInMethodsForEmail(auth, formData.email);
+    const emailExists = methods.length > 0;
+
+    if (emailExists) {
+      toast.warning("⚠️ Email already exists. Please login instead.");
+      setLoading(false);
+      return; // ⛔ STOP – no OTP will be sent
+    }
+
       // ✅ Generate 6-digit OTP
       const otp = Math.floor(100000 + Math.random() * 900000).toString();
       setGeneratedOtp(otp);
 
-      // ✅ Send OTP via EmailJS
+      // ✅ Send OTP
       await emailjs.send(
-        import.meta.env.VITE_EMAILJS_SERVICE_ID, // your EmailJS service ID
-        import.meta.env.VITE_EMAILJS_TEMPLATE_ID, // your EmailJS template ID
+        import.meta.env.VITE_EMAILJS_SERVICE_ID,
+        import.meta.env.VITE_EMAILJS_TEMPLATE_ID,
         {
-          email: formData.email, // matches {{email}}
-          passcode: otp, // matches {{passcode}}
-          time: new Date(Date.now() + 15 * 60000).toLocaleTimeString(), // matches {{time}}
+          email: formData.email,
+          passcode: otp,
+          time: new Date(Date.now() + 15 * 60000).toLocaleTimeString(),
         },
-        import.meta.env.VITE_EMAILJS_PUBLIC_KEY // your EmailJS public key
+        import.meta.env.VITE_EMAILJS_PUBLIC_KEY
       );
 
       toast.success("OTP sent to your email!");
-      setShowOtpPopup(true); // Show popup
+      setShowOtpPopup(true);
     } catch (error) {
+      setLoading(false);
       toast.error("Failed to send OTP.");
       console.error(error);
     } finally {
       setLoading(false);
     }
   };
+
   // ✅ OTP Verification & Registration
   const handleVerifyOtp = async () => {
     if (enteredOtp === generatedOtp) {
@@ -1234,7 +465,9 @@ const Register = ({ onClose, onSwitchToLogin }) => {
       {showOtpPopup && (
         <div className="fixed inset-0 flex items-center justify-center   z-50">
           <div className="bg-white p-6 rounded shadow-lg w-80">
-            <h3 className="text-lg font-semibold mb-4 text-gray-700">Verify Email</h3>
+            <h3 className="text-lg font-semibold mb-4 text-gray-700">
+              Verify Email
+            </h3>
             <input
               type="text"
               placeholder="Enter OTP"
